requirements_rl.txt
huggingface-hub[cli]>=0.19.2,<1.0
bitsandbytes>=0.43.0
black>=24.4.2
importlib-resources
matplotlib
packaging>=23.0
pycocotools
pycocoevalcap
pyyaml
regex
tqdm
timm
seaborn
scikit-learn
scipy
omegaconf
ipykernel
decord
tenacity
notebook
gradio
gradio-client
torch>=2.0
torchvision>=0.15.0
transformers==4.51.3
xformers
accelerate>=1.2.1
peft
lightning>=2.0.2
rouge
opencv-python-headless
prettytable
deepspeed==0.15.4
datasets==3.6.0
wandb==0.18.3
qwen_vl_utils==0.0.11
imagesize
colorama
vllm==0.7.3

distilabel[vllm,ray,openai]>=1.5.2
einops>=0.8.0
flake8>=6.0.0
hf_transfer>=0.1.4
isort>=5.12.0
liger_kernel==0.5.2
math-verify
parameterized>=0.9.0
pytest
safetensors>=0.3.3
sentencepiece>=0.1.99
trl==0.14.0
pillow
tensorboardx==2.6.2.2
latex2sympy2-extended
Levenshtein
pylatexenc
flash-attn==2.7.3